chore: moves search hooks to app folder
