+++ packages/core/src/experimental/index.ts
@@ -24,9 +24,9 @@ export { useProductGalleryQuery as useProductGalleryQuery_unstable } from '../..
 export { useProductLink as useProductLink_unstable } from '../../src/sdk/product/useProductLink'
 export { useProductQuery as useProductQuery_unstable } from '../../src/sdk/product/useProductQuery'
 export { useProductsPrefetch as useProductsPrefetch_unstable } from '../../src/sdk/product/useProductsPrefetch'
-export { default as useSearchHistory_unstable } from '../../src/sdk/search/useSearchHistory'
-export { default as useSuggestions_unstable } from '../../src/sdk/search/useSuggestions'
-export { default as useTopSearch_unstable } from '../../src/sdk/search/useTopSearch'
+export { default as useSearchHistory_unstable } from '../../app/sdk/search/useSearchHistory'
+export { default as useSuggestions_unstable } from '../../app/sdk/search/useSuggestions'
+export { default as useTopSearch_unstable } from '../../app/sdk/search/useTopSearch'
 export { useFilter as useFilter_unstable } from '../../src/sdk/search/useFilter'
 export { useDelayedFacets as useDelayedFacets_unstable } from '../../src/sdk/search/useDelayedFacets'
 export { useDelayedPagination as useDelayedPagination_unstable } from '../../src/sdk/search/useDelayedPagination'
